working ui and extra tests
