+++ tests/snapshots/test_load/test_esmfold/esmfold_6xHis_verts.txt
@@ -1,0 +1,60 @@
+-0.0887,0.2095,0.1076
+-0.085,0.1954,0.1063
+-0.0835,0.1915,0.0917
+-0.0954,0.1865,0.1131
+-0.0926,0.1937,0.0836
+-0.0949,0.187,0.128
+-0.1033,0.1924,0.1371
+-0.0846,0.1813,0.1353
+-0.0868,0.1833,0.1482
+-0.0981,0.19,0.1495
+-0.08,0.1981,0.0872
+-0.0806,0.1964,0.0727
+-0.0666,0.1955,0.0667
+-0.0884,0.2078,0.0662
+-0.0574,0.2024,0.0713
+-0.1032,0.2068,0.0682
+-0.1116,0.2138,0.0761
+-0.111,0.1976,0.0615
+-0.1236,0.199,0.0652
+-0.1243,0.2088,0.0741
+-0.0618,0.1877,0.0652
+-0.0504,0.1857,0.0563
+-0.0545,0.1871,0.0417
+-0.0441,0.1719,0.0587
+-0.0645,0.1814,0.0374
+-0.0345,0.1717,0.0702
+-0.0361,0.1673,0.0829
+-0.0216,0.1762,0.0692
+-0.0155,0.1747,0.0808
+-0.0241,0.1694,0.0894
+-0.0524,0.194,0.0338
+-0.0522,0.1965,0.0194
+-0.0414,0.1882,0.0125
+-0.0501,0.2114,0.0166
+-0.0304,0.1866,0.0178
+-0.0623,0.2197,0.0187
+-0.0654,0.2286,0.0284
+-0.073,0.2195,0.0101
+-0.0823,0.2278,0.0145
+-0.0779,0.2335,0.0256
+-0.0442,0.1782,0.0063
+-0.0358,0.1714,-0.0034
+-0.0391,0.1757,-0.0176
+-0.0372,0.1562,-0.002
+-0.0507,0.1784,-0.0208
+-0.0303,0.1506,0.01
+-0.0353,0.1454,0.0215
+-0.0166,0.1499,0.0111
+-0.0135,0.1445,0.0227
+-0.0246,0.1417,0.0292
+-0.0322,0.1801,-0.0257
+-0.0344,0.1782,-0.04
+-0.0383,0.1638,-0.0432
+-0.0219,0.1822,-0.0479
+-0.0336,0.1545,-0.0366
+-0.0203,0.197,-0.0495
+-0.0121,0.2058,-0.0432
+-0.0277,0.2044,-0.0585
+-0.0241,0.2171,-0.0577
+-0.0147,0.2182,-0.0485
