+++ tests/snapshots/test_load/test_md_load_gro_xtc/md_gro_xtc_verts.txt
@@ -1,0 +1,100 @@
+0.3918,0.2898,0.4846
+0.4054,0.2912,0.4796
+0.4065,0.3014,0.4681
+0.4043,0.3165,0.4716
+0.4195,0.302,0.4616
+0.4163,0.2932,0.4903
+0.4269,0.2863,0.4901
+0.4145,0.3025,0.4999
+0.424,0.3066,0.5104
+0.4162,0.3162,0.5207
+0.4245,0.3206,0.5324
+0.4105,0.329,0.5143
+0.4306,0.2942,0.5165
+0.4423,0.2951,0.5199
+0.423,0.2839,0.5197
+0.4284,0.2727,0.5274
+0.4181,0.2615,0.5293
+0.4227,0.2517,0.54
+0.4059,0.2657,0.5333
+0.4408,0.266,0.5206
+0.4513,0.2657,0.527
+0.4398,0.263,0.5074
+0.452,0.2578,0.5003
+0.4481,0.2534,0.4861
+0.4406,0.24,0.4852
+0.448,0.2283,0.4846
+0.4415,0.2162,0.4837
+0.4274,0.2162,0.4824
+0.4205,0.2048,0.4856
+0.4198,0.2282,0.4815
+0.4269,0.2402,0.4839
+0.4635,0.2682,0.5008
+0.4751,0.2646,0.5029
+0.4603,0.2815,0.5008
+0.4705,0.2918,0.5009
+0.4643,0.3057,0.498
+0.4737,0.3152,0.4899
+0.4527,0.305,0.4905
+0.4784,0.2921,0.514
+0.491,0.2939,0.5136
+0.4717,0.2916,0.5256
+0.4774,0.2914,0.5397
+0.4661,0.2923,0.5503
+0.4584,0.3058,0.5508
+0.4568,0.2817,0.5484
+0.4865,0.2791,0.5422
+0.4961,0.2803,0.55
+0.4832,0.2673,0.5365
+0.492,0.2556,0.5364
+0.4833,0.243,0.5345
+0.4729,0.2392,0.5454
+0.4649,0.2273,0.5405
+0.4802,0.2354,0.5582
+0.5042,0.2566,0.5269
+0.5151,0.2542,0.5315
+0.5018,0.2605,0.5143
+0.5121,0.2647,0.5048
+0.5054,0.268,0.491
+0.5143,0.2771,0.4813
+0.4994,0.2552,0.484
+0.4907,0.2581,0.472
+0.523,0.2747,0.5097
+0.535,0.2726,0.5097
+0.5179,0.2862,0.5149
+0.5266,0.2967,0.5202
+0.5191,0.3086,0.5253
+0.528,0.3205,0.5276
+0.5362,0.3228,0.5191
+0.5262,0.3297,0.5359
+0.5361,0.2906,0.5312
+0.5486,0.2913,0.5312
+0.5296,0.2842,0.5409
+0.5366,0.2777,0.5526
+0.5469,0.2673,0.5489
+0.5583,0.2685,0.5526
+0.5432,0.2582,0.5401
+0.552,0.2474,0.5346
+0.5427,0.2381,0.5258
+0.5345,0.228,0.534
+0.5254,0.2213,0.5241
+0.5428,0.2169,0.5411
+0.5635,0.2534,0.5257
+0.575,0.2489,0.5281
+0.5615,0.2635,0.5169
+0.5715,0.2698,0.5085
+0.5643,0.2795,0.4989
+0.5813,0.2777,0.5171
+0.5937,0.2773,0.5161
+0.576,0.2852,0.5271
+0.5839,0.292,0.538
+0.5755,0.3029,0.5452
+0.5738,0.3168,0.5385
+0.5617,0.3247,0.5443
+0.5605,0.3388,0.5394
+0.5535,0.3396,0.5256
+0.5911,0.283,0.5478
+0.6023,0.2867,0.552
+0.587,0.2701,0.549
+0.5941,0.2593,0.5555
+0.5846,0.2495,0.5618
